feat(service)t: 댓글 수정 및 삭제 시 작성자 검증 로직 추가
- CommentValidator를 사용하여 댓글 수정/삭제 시 작성자 검증 로직 추가
- 관련 테스트 케이스(updateComment_success, deleteComment_success) 수정 및 추가

+++ src/main/java/com/example/gunpo/service/board/CommentService.java
@@ -9,6 +9,7 @@ import com.example.gunpo.exception.board.CommentNotFoundException;
 import com.example.gunpo.repository.BoardRepository;
 import com.example.gunpo.repository.CommentRepository;
 import com.example.gunpo.service.member.AuthenticationService;
+import com.example.gunpo.validator.board.CommentValidator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,6 +22,7 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final BoardRepository boardRepository;
     private final AuthenticationService authenticationService;
+    private final CommentValidator commentValidator;
 
     @Transactional
     public Comment addComment(Long boardId, String accessToken, String content) {
@@ -37,14 +39,16 @@ public class CommentService {
     }
 
     @Transactional
-    public void updateComment(Long commentId, String content) {
+    public void updateComment(Long commentId, String accessToken,String content) {
         Comment comment = getComment(commentId);
+        commentValidator.verifyAuthor(accessToken,comment);
         comment.updateContent(content);
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId,String accessToken) {
         Comment comment = getComment(commentId);
+        commentValidator.verifyAuthor(accessToken,comment);
         commentRepository.delete(comment);
     }
 
@@ -60,6 +64,5 @@ public class CommentService {
         return commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommentNotFoundException(BoardErrorMessage.COMMENT_NOT_FOUND.getMessage()));
     }
-
 
 }
